Fix tests and refactoring
